+++ oneflow/core/operator/embedding_lookup_op.h
@@ -20,8 +20,8 @@ class EmbeddingLookupOp final : public Operator {
   int32_t MaxModelSplitNum() const override { return op_conf().embedding_lookup_conf().units(); }
 
  private:
-  void InferOutBlobModelSplitAxis(std::function<int64_t*(const std::string&)> ModelSplitAxis4BnInOp,
-                                  std::function<int64_t(const std::string&)> ShapeNumAxes4BnInOp,
+  void InferOutBlobModelSplitAxis(std::function<int32_t*(const std::string&)> ModelSplitAxis4BnInOp,
+                                  std::function<int32_t(const std::string&)> ShapeNumAxes4BnInOp,
                                   const ParallelContext* parallel_context) const override {
     NaiveInferOutBlobModelSplitAxis(ModelSplitAxis4BnInOp, ShapeNumAxes4BnInOp, parallel_context);
   }
