+++ oneflow/core/operator/print_op.h
@@ -20,8 +20,8 @@ class PrintOp final : public Operator {
                       const ParallelContext* parallel_ctx) const override {}
 
  private:
-  void InferOutBlobModelSplitAxis(std::function<int64_t*(const std::string&)> ModelSplitAxis4BnInOp,
-                                  std::function<int64_t(const std::string&)> ShapeNumAxes4BnInOp,
+  void InferOutBlobModelSplitAxis(std::function<int32_t*(const std::string&)> ModelSplitAxis4BnInOp,
+                                  std::function<int32_t(const std::string&)> ShapeNumAxes4BnInOp,
                                   const ParallelContext* parallel_context) const override {}
 
   LogicalBlobId ibn2lbi(const std::string& input_bn) const override;
